Migrate debug menu to new initialization approach
