+++ modules/app/src/debug/kotlin/kekmech/ru/mpeiapp/FeatureDebugMenuModuleAccessor.kt
@@ -1,6 +1,6 @@
 package kekmech.ru.mpeiapp
 
-import kekmech.ru.feature_debug_menu.di.FeatureDebugMenuModule
+import kekmech.ru.debug_menu.di.FeatureDebugMenuModule
 import org.koin.core.KoinApplication
 
 fun KoinApplication.includeFeatureDebugMenuModule() {
