+++ modules/debug_menu/src/main/AndroidManifest.xml
@@ -4,11 +4,11 @@
 
     <application>
         <service
-            android:name="kekmech.ru.feature_debug_menu.presentation.DebugMenuService"
+            android:name="kekmech.ru.debug_menu.presentation.DebugMenuService"
             android:exported="false"
             android:stopWithTask="true" />
         <activity
-            android:name="kekmech.ru.feature_debug_menu.presentation.DebugMenuActivity"
+            android:name="kekmech.ru.debug_menu.presentation.DebugMenuActivity"
             android:exported="false"
             android:label="MpeiX Debug Menu"
             android:theme="@style/Theme.Mpeix" />
