+++ modules/debug_menu/src/main/kotlin/kekmech/ru/debug_menu/presentation/screens/main/DebugMenuScreen.kt
@@ -1,13 +1,13 @@
-package kekmech.ru.feature_debug_menu.presentation.screens.main
+package kekmech.ru.debug_menu.presentation.screens.main
 
 import androidx.compose.runtime.Composable
 import com.bumble.appyx.core.modality.BuildContext
 import com.bumble.appyx.core.node.Node
 import kekmech.ru.common_elm.elmNode
 import kekmech.ru.common_navigation_api.NavTarget
-import kekmech.ru.feature_debug_menu.presentation.screens.main.elm.DebugMenuState
-import kekmech.ru.feature_debug_menu.presentation.screens.main.elm.DebugMenuStore
-import kekmech.ru.feature_debug_menu.presentation.screens.main.elm.DebugMenuStoreFactory
+import kekmech.ru.debug_menu.presentation.screens.main.elm.DebugMenuState
+import kekmech.ru.debug_menu.presentation.screens.main.elm.DebugMenuStore
+import kekmech.ru.debug_menu.presentation.screens.main.elm.DebugMenuStoreFactory
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
