+++ modules/debug_menu/src/main/kotlin/kekmech/ru/debug_menu/presentation/screens/main/elm/DebugMenuActor.kt
@@ -1,0 +1,15 @@
+package kekmech.ru.debug_menu.presentation.screens.main.elm
+
+import kekmech.ru.debug_menu.presentation.screens.main.elm.DebugMenuCommand
+import kekmech.ru.debug_menu.presentation.screens.main.elm.DebugMenuEvent
+import kotlinx.coroutines.flow.Flow
+import vivid.money.elmslie.coroutines.Actor
+import kekmech.ru.debug_menu.presentation.screens.main.elm.DebugMenuCommand as Command
+import kekmech.ru.debug_menu.presentation.screens.main.elm.DebugMenuEvent as Event
+
+internal class DebugMenuActor : Actor<Command, Event> {
+
+    override fun execute(command: DebugMenuCommand): Flow<DebugMenuEvent> {
+        TODO("Not yet implemented")
+    }
+}
