+++ modules/debug_menu/src/main/res/drawable/ic_debug_menu_24.xml
