--- modules/feature/debug_menu/src/main/kotlin/kekmech/ru/feature_debug_menu/di/FeatureDebugMenuModule.kt
@@ -1,7 +1,0 @@
-package kekmech.ru.feature_debug_menu.di
-
-import org.koin.dsl.module
-
-val FeatureDebugMenuModule = module {
-
-}
